src/requirements.txt
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
exceptiongroup==1.1.3
fastapi==0.68.0
Flask==2.3.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pydantic==1.10.12
PySocks==1.7.1
pytest==7.4.2
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.14.2
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.15.0
Werkzeug==2.3.7
wsproto==1.2.0
